+++ packages/trpc/package.json
@@ -15,6 +15,7 @@
     "@trpc/server": "10.9.0",
     "firebase-admin": "11.5.0",
     "otplib": "12.0.1",
+    "request-ip": "3.3.0",
     "zod": "3.20.2"
   },
   "devDependencies": {
@@ -29,8 +30,8 @@
     "eslint-config-turbo": "0.0.7",
     "eslint-import-resolver-typescript": "3.5.3",
     "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-jsx-a11y": "6.7.1",
+    "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-react": "7.32.1",
     "eslint-plugin-react-hooks": "4.6.0",
     "prettier": "2.8.3",
