+++ packages/trpc/src/trpc.ts
@@ -1,5 +1,8 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
 import { initTRPC } from '@trpc/server';
 import { type CreateNextContextOptions } from '@trpc/server/adapters/next';
+// @ts-ignore
+import requestIp from 'request-ip';
 
 type CreateContextOptions = Record<string, never>;
 
@@ -8,6 +11,22 @@ export const createInnerTRPCContext = async (_opts: CreateContextOptions) => {
 };
 
 export const createTRPCContext = async (_opts: CreateNextContextOptions) => {
+  const { req } = _opts;
+  const detectedIp = requestIp.getClientIp(req);
+  console.log('ip', detectedIp);
+  // @ts-ignore
+  const forwarded = req.headers['x-forwarded-for'];
+  console.log('forwarded', forwarded);
+  // @ts-ignore
+  console.log('remoteAddress', req?.socket?.remoteAddress);
+  console.log('headers', req.headers);
+  console.log('headers str', JSON.stringify(req.headers));
+  const compIp =
+    typeof forwarded === 'string'
+      ? forwarded.split(/, /)[0]
+      : // @ts-ignore
+        req?.socket?.remoteAddress;
+  console.log('compIp', compIp);
   return createInnerTRPCContext({});
 };
